Moved decorator to utilities, removed incorrect decorator in questions and moved to ask
